CENTRALIZED V1/centralised_website/requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
requests
djangorestframework-simplejwt